system-probe: move kernel out helper to common rspecs
