+++ test/kitchen/test/integration/common/rspec/kernel_out_spec_helper.rb
