+++ test/kitchen/test/integration/system-probe-test/rspec/kernel_out_spec_helper.rb
@@ -1,0 +1,1 @@
+../../common/rspec/kernel_out_spec_helper.rb
